backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
pydantic==2.6.1

# Audio Processing
librosa==0.10.1
numpy==1.26.4
soundfile==0.12.1
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
requests==2.31.0

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0